--- a/vm/src/number.rs
+++ b/vm/src/number.rs
@@ -2,19 +2,23 @@ use crate::{value::Value, Error};
 use core::fmt::{self, Display, Formatter};
 
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub struct Number(i64);
+pub struct Number(u64);
 
 impl Number {
     pub const fn new(number: i64) -> Self {
-        Self(number)
+        Self(number as u64)
     }
 
     pub const fn to_i64(self) -> i64 {
-        self.0
+        self.0 as i64 >> 1
+    }
+
+    pub const fn from_raw(raw: u64) -> Self {
+        Self(raw)
     }
 
     pub const fn to_raw(self) -> u64 {
-        self.0 as u64
+        self.0
     }
 }