Before: 

```sh
+ hyperfine --sort command 'target/release/stak bench/fibonacci/main.out' 'gsi bench/fibonacci/main.scm' 'python3 bench/fibonacci/main.py' 'csi -s bench/fibonacci/main.scm'
Benchmark 1: target/release/stak bench/fibonacci/main.out
  Time (mean ± σ):     12.780 s ±  0.316 s    [User: 12.662 s, System: 0.029 s]
  Range (min … max):   12.639 s … 13.668 s    10 runs

  Warning: The first benchmarking run for this command was significantly slower than the rest (13.668 s). This could be caused by (filesystem) caches that were not filled until after the first run. You should consider using the '--warmup' option to fill those caches before the actual benchmark. Alternatively, use the '--prepare' option to clear the caches before each timing run.

Benchmark 2: gsi bench/fibonacci/main.scm
  Time (mean ± σ):     10.605 s ±  0.064 s    [User: 10.566 s, System: 0.035 s]
  Range (min … max):   10.565 s … 10.786 s    10 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.

Benchmark 3: python3 bench/fibonacci/main.py
  Time (mean ± σ):      4.077 s ±  0.016 s    [User: 4.064 s, System: 0.011 s]
  Range (min … max):    4.060 s …  4.104 s    10 runs

Benchmark 4: csi -s bench/fibonacci/main.scm
  Time (mean ± σ):     21.159 s ±  0.380 s    [User: 20.962 s, System: 0.176 s]
  Range (min … max):   20.654 s … 21.743 s    10 runs

Relative speed comparison
        3.13 ±  0.08  target/release/stak bench/fibonacci/main.out
        2.60 ±  0.02  gsi bench/fibonacci/main.scm
        1.00          python3 bench/fibonacci/main.py
        5.19 ±  0.10  csi -s bench/fibonacci/main.scm
+ for file in '$(find bench -type f -name '\''*.scm'\'' | sort)'
+ base=bench/sum/main
+ cat prelude.scm bench/sum/main.scm
+ ./main.scm
+ hyperfine --sort command 'target/release/stak bench/sum/main.out' 'gsi bench/sum/main.scm' 'python3 bench/sum/main.py' 'csi -s bench/sum/main.scm'
Benchmark 1: target/release/stak bench/sum/main.out
  Time (mean ± σ):      9.989 s ±  0.032 s    [User: 9.956 s, System: 0.030 s]
  Range (min … max):    9.960 s … 10.066 s    10 runs

Benchmark 2: gsi bench/sum/main.scm
  Time (mean ± σ):      9.112 s ±  0.046 s    [User: 9.062 s, System: 0.038 s]
  Range (min … max):    9.057 s …  9.169 s    10 runs

Benchmark 3: python3 bench/sum/main.py
  Time (mean ± σ):      2.427 s ±  0.031 s    [User: 2.413 s, System: 0.013 s]
  Range (min … max):    2.388 s …  2.470 s    10 runs

Benchmark 4: csi -s bench/sum/main.scm
  Time (mean ± σ):     19.684 s ±  0.149 s    [User: 19.524 s, System: 0.157 s]
  Range (min … max):   19.351 s … 19.846 s    10 runs

Relative speed comparison
        4.12 ±  0.05  target/release/stak bench/sum/main.out
        3.75 ±  0.05  gsi bench/sum/main.scm
        1.00          python3 bench/sum/main.py
        8.11 ±  0.12  csi -s bench/sum/main.scm
+ for file in '$(find bench -type f -name '\''*.scm'\'' | sort)'
+ base=bench/tak/main
+ cat prelude.scm bench/tak/main.scm
+ ./main.scm
+ hyperfine --sort command 'target/release/stak bench/tak/main.out' 'gsi bench/tak/main.scm' 'python3 bench/tak/main.py' 'csi -s bench/tak/main.scm'
Benchmark 1: target/release/stak bench/tak/main.out
  Time (mean ± σ):      5.238 s ±  0.008 s    [User: 5.223 s, System: 0.014 s]
  Range (min … max):    5.227 s …  5.248 s    10 runs

Benchmark 2: gsi bench/tak/main.scm
  Time (mean ± σ):      3.886 s ±  0.022 s    [User: 3.866 s, System: 0.018 s]
  Range (min … max):    3.868 s …  3.929 s    10 runs

Benchmark 3: python3 bench/tak/main.py
  Time (mean ± σ):      1.479 s ±  0.024 s    [User: 1.470 s, System: 0.008 s]
  Range (min … max):    1.453 s …  1.533 s    10 runs

Benchmark 4: csi -s bench/tak/main.scm
  Time (mean ± σ):      7.727 s ±  0.160 s    [User: 7.647 s, System: 0.078 s]
  Range (min … max):    7.515 s …  7.975 s    10 runs

Relative speed comparison
        3.54 ±  0.06  target/release/stak bench/tak/main.out
        2.63 ±  0.04  gsi bench/tak/main.scm
        1.00          python3 bench/tak/main.py
        5.22 ±  0.14  csi -s bench/tak/main.scm
```

After:

```sh
+ hyperfine --sort command 'target/release/stak bench/fibonacci/main.out' 'gsi bench/fibonacci/main.scm' 'python3 bench/fibonacci/main.py' 'csi -s bench/fibonacci/main.scm'
Benchmark 1: target/release/stak bench/fibonacci/main.out
  Time (mean ± σ):     11.814 s ±  0.371 s    [User: 11.665 s, System: 0.037 s]
  Range (min … max):   11.656 s … 12.848 s    10 runs

  Warning: The first benchmarking run for this command was significantly slower than the rest (12.848 s). This could be caused by (filesystem) caches that were not filled until after the first run. You should consider using the '--warmup' option to fill those caches before the actual benchmark. Alternatively, use the '--prepare' option to clear the caches before each timing run.

Benchmark 2: gsi bench/fibonacci/main.scm
  Time (mean ± σ):     10.611 s ±  0.057 s    [User: 10.566 s, System: 0.041 s]
  Range (min … max):   10.563 s … 10.762 s    10 runs

Benchmark 3: python3 bench/fibonacci/main.py
  Time (mean ± σ):      4.097 s ±  0.027 s    [User: 4.080 s, System: 0.014 s]
  Range (min … max):    4.069 s …  4.149 s    10 runs

Benchmark 4: csi -s bench/fibonacci/main.scm
  Time (mean ± σ):     21.125 s ±  0.330 s    [User: 20.938 s, System: 0.169 s]
  Range (min … max):   20.642 s … 21.633 s    10 runs

Relative speed comparison
        2.88 ±  0.09  target/release/stak bench/fibonacci/main.out
        2.59 ±  0.02  gsi bench/fibonacci/main.scm
        1.00          python3 bench/fibonacci/main.py
        5.16 ±  0.09  csi -s bench/fibonacci/main.scm
+ for file in '$(find bench -type f -name '\''*.scm'\'' | sort)'
+ base=bench/sum/main
+ cat prelude.scm bench/sum/main.scm
+ ./main.scm
+ hyperfine --sort command 'target/release/stak bench/sum/main.out' 'gsi bench/sum/main.scm' 'python3 bench/sum/main.py' 'csi -s bench/sum/main.scm'
Benchmark 1: target/release/stak bench/sum/main.out
  Time (mean ± σ):      9.240 s ±  0.093 s    [User: 9.214 s, System: 0.020 s]
  Range (min … max):    9.191 s …  9.494 s    10 runs

  Warning: The first benchmarking run for this command was significantly slower than the rest (9.494 s). This could be caused by (filesystem) caches that were not filled until after the first run. You should consider using the '--warmup' option to fill those caches before the actual benchmark. Alternatively, use the '--prepare' option to clear the caches before each timing run.

Benchmark 2: gsi bench/sum/main.scm
  Time (mean ± σ):      9.091 s ±  0.042 s    [User: 9.052 s, System: 0.030 s]
  Range (min … max):    9.053 s …  9.172 s    10 runs

Benchmark 3: python3 bench/sum/main.py
  Time (mean ± σ):      2.378 s ±  0.030 s    [User: 2.363 s, System: 0.012 s]
  Range (min … max):    2.338 s …  2.440 s    10 runs

Benchmark 4: csi -s bench/sum/main.scm
  Time (mean ± σ):     19.683 s ±  0.167 s    [User: 19.515 s, System: 0.163 s]
  Range (min … max):   19.254 s … 19.859 s    10 runs

Relative speed comparison
        3.89 ±  0.06  target/release/stak bench/sum/main.out
        3.82 ±  0.05  gsi bench/sum/main.scm
        1.00          python3 bench/sum/main.py
        8.28 ±  0.13  csi -s bench/sum/main.scm
+ for file in '$(find bench -type f -name '\''*.scm'\'' | sort)'
+ base=bench/tak/main
+ cat prelude.scm bench/tak/main.scm
+ ./main.scm
+ hyperfine --sort command 'target/release/stak bench/tak/main.out' 'gsi bench/tak/main.scm' 'python3 bench/tak/main.py' 'csi -s bench/tak/main.scm'
Benchmark 1: target/release/stak bench/tak/main.out
  Time (mean ± σ):      4.753 s ±  0.074 s    [User: 4.742 s, System: 0.010 s]
  Range (min … max):    4.690 s …  4.867 s    10 runs

  Warning: The first benchmarking run for this command was significantly slower than the rest (4.867 s). This could be caused by (filesystem) caches that were not filled until after the first run. You should consider using the '--warmup' option to fill those caches before the actual benchmark. Alternatively, use the '--prepare' option to clear the caches before each timing run.

Benchmark 2: gsi bench/tak/main.scm
  Time (mean ± σ):      3.872 s ±  0.007 s    [User: 3.854 s, System: 0.016 s]
  Range (min … max):    3.860 s …  3.881 s    10 runs

Benchmark 3: python3 bench/tak/main.py
  Time (mean ± σ):      1.455 s ±  0.011 s    [User: 1.447 s, System: 0.006 s]
  Range (min … max):    1.437 s …  1.470 s    10 runs

Benchmark 4: csi -s bench/tak/main.scm
  Time (mean ± σ):      7.797 s ±  0.093 s    [User: 7.732 s, System: 0.064 s]
  Range (min … max):    7.652 s …  7.963 s    10 runs

Relative speed comparison
        3.27 ±  0.06  target/release/stak bench/tak/main.out
        2.66 ±  0.02  gsi bench/tak/main.scm
        1.00          python3 bench/tak/main.py
        5.36 ±  0.08  csi -s bench/tak/main.scm
```